feat: addressing review comments
