--- packages/amplify-provider-awscloudformation/src/aws-utils/aws-cfn-progress-formatter.js
@@ -1,45 +1,0 @@
-// disabling lint until this file is converted to TS
-/* eslint-disable */
-const chalk = require('chalk');
-const columnify = require('columnify');
-
-const COLUMNIFY_WIDTH = 20;
-
-const CFN_SUCCESS_STATUS = ['UPDATE_COMPLETE', 'CREATE_COMPLETE', 'DELETE_COMPLETE', 'DELETE_SKIPPED'];
-const CNF_ERROR_STATUS = ['CREATE_FAILED', 'DELETE_FAILED', 'UPDATE_FAILED'];
-
-// Custom Item formatter for progressbar
-const createItemFormatter = (payload) => {
-  const e = [{
-    logicalResourceId: payload.LogicalResourceId,
-    resourceType: payload.ResourceType,
-    resourceStatus: payload.ResourceStatus,
-    timeStamp: (new Date(payload.Timestamp)).toString(),
-  }];
-
-  let output = columnify(e, {
-    showHeaders: false,
-    truncate: true,
-    maxWidth: COLUMNIFY_WIDTH,
-    minWidth: COLUMNIFY_WIDTH,
-  });
-
-  if(CFN_SUCCESS_STATUS.includes(payload.ResourceStatus)) {
-    output = chalk.green(output);
-  }
-  if (CNF_ERROR_STATUS.includes(payload.ResourceStatus)) {
-    output = chalk.red(output);
-  }
-  return output;
-};
-
-// Custom progress bar formatter
-const createProgressBarFormatter = (payload) => {
-  const progressNameParts = payload.progressName.split('-');
-  const name = progressNameParts.length === 1 ? progressNameParts[0] : `${progressNameParts[0]} ${progressNameParts[1]}`;
-  return `Deploying ${name} on env: ${payload.envName}`;
-};
-module.exports = {
-  createItemFormatter,
-  createProgressBarFormatter
-}
